Moving test extension to SomeExtension
