+++ tests/test-extensions/some-extension/src/main/java/org/mule/test/some/extension/SomeExtension.java
@@ -10,7 +10,9 @@ package org.mule.test.some.extension;
 import org.mule.runtime.extension.api.annotation.Export;
 import org.mule.runtime.extension.api.annotation.Extension;
 import org.mule.runtime.extension.api.annotation.Operations;
+import org.mule.runtime.extension.api.annotation.Sources;
 import org.mule.runtime.extension.api.annotation.connectivity.ConnectionProviders;
+import org.mule.runtime.extension.api.annotation.dsl.xml.Xml;
 import org.mule.runtime.extension.api.annotation.error.ErrorTypes;
 import org.mule.test.heisenberg.extension.HeisenbergErrors;
 
@@ -19,6 +21,8 @@ import org.mule.test.heisenberg.extension.HeisenbergErrors;
 @ErrorTypes(HeisenbergErrors.class)
 @Operations(SomeOps.class)
 @Export(classes = CustomConnectionException.class)
+@Sources({SomeEmittingSource.class})
+@Xml(namespace = "http://www.mulesoft.org/schema/mule/some", prefix = "some")
 public class SomeExtension {
 
 }
