+++ tests/test-policies/pom.xml
@@ -71,7 +71,7 @@
         <!-- Required test extensions -->
         <dependency>
             <groupId>org.mule.tests</groupId>
-            <artifactId>mule-values-extension</artifactId>
+            <artifactId>mule-some-extension</artifactId>
             <version>${project.version}</version>
             <classifier>mule-plugin</classifier>
             <scope>test</scope>
