+++ tests/test-policies/src/test/resources/test-policy-config.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <mule xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
       xmlns:test-policy="http://www.mulesoft.org/schema/mule/test-policy"
-      xmlns:values="http://www.mulesoft.org/schema/mule/values"
+      xmlns:some="http://www.mulesoft.org/schema/mule/some"
       xmlns="http://www.mulesoft.org/schema/mule/core"
       xsi:schemaLocation="http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd
                http://www.mulesoft.org/schema/mule/test-policy http://www.mulesoft.org/schema/mule/test-policy/current/mule-test-policy.xsd
-               http://www.mulesoft.org/schema/mule/values http://www.mulesoft.org/schema/mule/values/current/mule-values.xsd">
+               http://www.mulesoft.org/schema/mule/some http://www.mulesoft.org/schema/mule/some/current/mule-some.xsd">
 
     <object name="disposeListener" class="org.mule.test.policy.PolicyTestCase$DisposeListenerMessageProcessor"/>
 
     <flow name="main">
-        <values:value-emitting-source message="hello dog!" times="1"/>
+        <some:some-emitting-source message="hello dog!" times="1"/>
         <logger level="ERROR" message="#[payload]"/>
         <processor ref="disposeListener"/>
     </flow>
